[✨ Feature] 일정 추가, 삭제 폼 유효성 검사, 스타일링 수정 (#60)
* fix: 인터페이스 타입 관련 오류 수정 및 회원가입 커스텀 훅 코드 변경

* refactor: 수정 db 구조에 맞게 비동기 호출 함수 수정 (#51)

* style: 캘린더 주말 색 변경 (#51)

* style: 일정 관리 페이지  설정 해상도 반영 (#51)

* feat:  일정 추가, 수정 폼 유효성 검사, 스타일링 수정 (#51)

* fix: merge 충돌 해결

* refactor: 스케줄 필드 이름 수정 (#51)

* fix: hasError prop 이름 수정 (#51)

* refactor: user 정보 redux로 관리 (#51)

---------

+++ src/redux/actions/scheduleActions.ts
@@ -1,15 +1,7 @@
-import {
-	doc,
-	setDoc,
-	collection,
-	updateDoc,
-	deleteDoc,
-	getDoc,
-	Timestamp,
-} from 'firebase/firestore';
+import { doc, setDoc, getDoc, updateDoc, Timestamp } from 'firebase/firestore';
 import { db } from '@/firebaseConfig';
 import { AppThunk } from '@/redux/store';
-import { scheduleApiResponse, TSchedule } from '@/types/schedule';
+import { TScheduleApiResponse, TSchedule } from '@/types/schedule';
 import {
 	GET_SCHEDULES,
 	ADD_SCHEDULES,
@@ -18,11 +10,18 @@ import {
 	SELECT_DATE,
 	FILTERED_SCHEDULES,
 	SET_LOADING,
+	SET_SCHEDULE_MODAL_OPEN,
 } from '../actionTypes';
 
-const setisLoading = (isLoading: boolean) => ({
+export const setisLoading = (isLoading: boolean) => ({
 	type: SET_LOADING,
 	payload: isLoading,
+});
+
+//스케줄 추가, 수정 모달 상태
+export const setIsScheduleModalOpen = (isScheduleModalOpen: boolean) => ({
+	type: SET_SCHEDULE_MODAL_OPEN,
+	payload: isScheduleModalOpen,
 });
 
 export const getSchedules = (schedules: TSchedule[]) => ({
@@ -45,7 +44,7 @@ const removeSchedules = (scheduleIds: string[]) => ({
 	payload: scheduleIds,
 });
 
-export const selectDate = (date: Date | null) => ({
+export const selectDate = (date: Date | Timestamp) => ({
 	type: SELECT_DATE,
 	payload: date,
 });
@@ -62,38 +61,47 @@ export const filteredSchedules = (schedules: TSchedule[]) => ({
  *          ├─ schedule_id_1 (문서)
  *          ├─ schedule_id_2 (문서)
  */
-// firestore에 스케줄 추가
+// Firestore에 스케줄 추가
 export const addScheduleToFirestore = (
-	userId: string,
+	userId: string | undefined,
 	schedules: TSchedule[],
-): AppThunk<Promise<scheduleApiResponse<void>>> => {
+): AppThunk<Promise<TScheduleApiResponse<void>>> => {
 	return async (dispatch) => {
-		dispatch(setisLoading(true)); // 로딩 시작
+		if (!userId) {
+			return {
+				success: false,
+				message: '사용자 ID가 없습니다.',
+			};
+		}
+
+		dispatch(setisLoading(true));
 		try {
-			const schedulesCollectionRef = collection(db, 'schedules', userId, 'userSchedules');
-			const addPromises = schedules.map(async (schedule) => {
-				const scheduleDocRef = doc(schedulesCollectionRef, schedule.schedule_id); // schedule_id로 문서 생성
-				// Firestore에서 schedule_id 확인
-				const scheduleSnapshot = await getDoc(scheduleDocRef);
-				if (!scheduleSnapshot.exists()) {
-					const newSchedule = {
+			const userDocRef = doc(db, 'schedules', userId);
+			const userDoc = await getDoc(userDocRef);
+
+			if (!userDoc.exists()) {
+				// 사용자 문서가 없으면 새로 생성
+				await setDoc(userDocRef, {
+					schedules: schedules.map((schedule) => ({
 						...schedule,
-						created_at: Timestamp.now(), // 현재 시간을 Firestore Timestamp로 저장
-					};
-					return setDoc(scheduleDocRef, newSchedule);
-				} else {
-					// 존재하는 경우 로그 출력
-					console.log(`스케줄 ${schedule.schedule_id} 가 이미 존재함`);
-					return Promise.resolve(); // 이미 존재하는 경우에도 Promise 반환
-				}
-			});
+						created_at: Timestamp.now(),
+					})),
+				});
+			} else {
+				// 기존 스케줄 배열에 새 스케줄 추가
+				const existingSchedules = userDoc.data().schedules || [];
+				const updatedSchedules = [
+					...existingSchedules,
+					...schedules.map((schedule) => ({
+						...schedule,
+						created_at: Timestamp.now(),
+					})),
+				];
 
-			// 모든 추가 작업이 완료될 때까지 기다림
-			await Promise.all(addPromises);
+				await updateDoc(userDocRef, { schedules: updatedSchedules });
+			}
 
-			// Redux 상태 업데이트
 			dispatch(addSchedules(schedules));
-
 			return {
 				success: true,
 				message: '일정을 성공적으로 추가했습니다.',
@@ -105,38 +113,44 @@ export const addScheduleToFirestore = (
 				message: '일정 추가 중 오류가 발생했습니다.',
 			};
 		} finally {
-			dispatch(setisLoading(false)); // 로딩 종료
+			dispatch(setisLoading(false));
 		}
 	};
 };
 
-// firestore에 스케줄 수정
+// Firestore에서 스케줄 수정
 export const editScheduleToFirestore = (
-	userId: string,
-	schedules: TSchedule[],
-): AppThunk<Promise<scheduleApiResponse<void>>> => {
+	userId: string | undefined,
+	updatedSchedules: TSchedule[],
+): AppThunk<Promise<TScheduleApiResponse<void>>> => {
 	return async (dispatch) => {
-		dispatch(setisLoading(true)); // 로딩 시작
-		try {
-			const editPromises = schedules.map(async (schedule) => {
-				// userSchedules 컬렉션의 문서 참조
-				const scheduleDocRef = doc(db, 'schedules', userId, 'userSchedules', schedule.schedule_id);
-				const scheduleSnapshot = await getDoc(scheduleDocRef);
+		if (!userId) {
+			return {
+				success: false,
+				message: '사용자 ID가 없습니다.',
+			};
+		}
 
-				if (scheduleSnapshot.exists()) {
-					// Firestore에서 문서가 존재하면 수정
-					return updateDoc(scheduleDocRef, { ...schedule });
-				} else {
-					console.log(`${schedule.schedule_id} 가 존재하지 않음`);
-					return Promise.resolve(); // 존재하지 않는 경우에도 Promise를 반환
-				}
+		dispatch(setisLoading(true));
+		try {
+			const userDocRef = doc(db, 'schedules', userId);
+			const userDoc = await getDoc(userDocRef);
+
+			if (!userDoc.exists()) {
+				throw new Error('사용자 문서가 존재하지 않습니다.');
+			}
+
+			const existingSchedules: TSchedule[] = userDoc.data().schedules || [];
+			const updatedSchedulesList = existingSchedules.map((schedule) => {
+				const updatedSchedule = updatedSchedules.find(
+					(s) => s.schedule_id === schedule.schedule_id,
+				);
+				return updatedSchedule || schedule;
 			});
 
-			// 모든 수정 작업이 완료될 때까지 기다림
-			await Promise.all(editPromises);
+			await updateDoc(userDocRef, { schedules: updatedSchedulesList });
+			dispatch(editSchedules(updatedSchedules));
 
-			// Redux 상태 업데이트
-			dispatch(editSchedules(schedules));
 			return {
 				success: true,
 				message: '일정을 성공적으로 수정했습니다.',
@@ -148,37 +162,40 @@ export const editScheduleToFirestore = (
 				message: '일정 수정 중 오류가 발생했습니다.',
 			};
 		} finally {
-			dispatch(setisLoading(false)); // 로딩 종료
+			dispatch(setisLoading(false));
 		}
 	};
 };
 
-// firestore에 스케줄 삭제
+// Firestore에서 스케줄 삭제
 export const removeScheduleToFirestore = (
-	userId: string,
+	userId: string | undefined,
 	scheduleIds: string[],
-): AppThunk<Promise<scheduleApiResponse<void>>> => {
+): AppThunk<Promise<TScheduleApiResponse<void>>> => {
 	return async (dispatch) => {
-		dispatch(setisLoading(true)); // 로딩 시작
+		if (!userId) {
+			return {
+				success: false,
+				message: '사용자 ID가 없습니다.',
+			};
+		}
+
+		dispatch(setisLoading(true));
 		try {
-			const deletePromises = scheduleIds.map(async (scheduleId) => {
-				// userSchedules 컬렉션의 문서 참조
-				const scheduleDocRef = doc(db, 'schedules', userId, 'userSchedules', scheduleId);
-				const scheduleSnapshot = await getDoc(scheduleDocRef);
+			const userDocRef = doc(db, 'schedules', userId);
+			const userDoc = await getDoc(userDocRef);
 
-				if (scheduleSnapshot.exists()) {
-					// Firestore에서 문서가 존재하면 삭제
-					return deleteDoc(scheduleDocRef);
-				} else {
-					console.log(`${scheduleId} 가 존재하지 않음`);
-					return Promise.resolve(); // 존재하지 않는 경우에도 Promise를 반환
-				}
-			});
+			if (!userDoc.exists()) {
+				throw new Error('사용자 문서가 존재하지 않습니다.');
+			}
 
-			// 모든 삭제 작업이 완료될 때까지 기다림
-			await Promise.all(deletePromises);
+			const existingSchedules: TSchedule[] = userDoc.data().schedules || [];
 
-			// Redux 상태 업데이트
+			const filteredSchedules = existingSchedules.filter(
+				(schedule) => !scheduleIds.includes(schedule.schedule_id),
+			);
+
+			await updateDoc(userDocRef, { schedules: filteredSchedules });
 			dispatch(removeSchedules(scheduleIds));
 
 			return {
@@ -192,7 +209,7 @@ export const removeScheduleToFirestore = (
 				message: '일정 삭제 중 오류가 발생했습니다.',
 			};
 		} finally {
-			dispatch(setisLoading(false)); // 로딩 종료
+			dispatch(setisLoading(false));
 		}
 	};
 };
